# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Asisteam HTTP",
    "version": "1.0.0",
    "description": "MIG-07. Grupos/perfil implementados con sesión temporal y SQL canónico. Refines Zod (edad/roles/URLs) y reglas SQL no se sustituyen por JSON Schema."
  },
  "paths": {
...
    "/api/v1/me/groups": {
      "get": {
        "operationId": "listMyGroups",
        "summary": "Grupos con membresía ACTIVE; roles locales unidos, orden name/id",
        "tags": [
          "groups"
        ],
        "x-implementation-status": "implemented",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 50
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MyGroups"
                }
              }
            }
          },
          "400": {
            "description": "Error 400; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error 401; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error 403; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Error 404; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Error 409; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Error 422; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error 429; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error 500; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error 503; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Error 504; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}": {
      "get": {
        "operationId": "getGroup",
        "summary": "Detalle visible; código/settings solo ADMIN, grupo ajeno 404",
        "tags": [
          "groups"
        ],
        "x-implementation-status": "implemented",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respuesta válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupDetail"
                }
              }
            }
          },
          "400": {
            "description": "Error 400; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error 401; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error 403; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Error 404; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Error 409; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Error 422; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error 429; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error 500; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error 503; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Error 504; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateGroup",
        "summary": "Actualizar datos del grupo; ADMIN ACTIVE y RLS",
        "tags": [
          "groups"
        ],
        "x-implementation-status": "implemented",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "400": {
            "description": "Error 400; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error 401; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error 403; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Error 404; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Error 409; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Error 422; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error 429; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error 500; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error 503; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Error 504; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups": {
      "post": {
        "operationId": "createGroup",
        "summary": "Crear grupo + ADMIN mediante create_group; actor de sesión",
        "tags": [
          "groups"
        ],
        "x-implementation-status": "implemented",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroup"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Respuesta válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupCreated"
                }
              }
            }
          },
          "400": {
            "description": "Error 400; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error 401; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error 403; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Error 404; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Error 409; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Error 422; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error 429; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error 500; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error 503; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Error 504; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/groups/{groupId}/settings": {
      "patch": {
        "operationId": "updateGroupSettings",
        "summary": "Toggles independientes mediante RPC",
        "tags": [
          "groups"
        ],
        "x-implementation-status": "implemented",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GroupSettingsChange"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Respuesta válida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupSettings"
                }
              }
            }
          },
          "400": {
            "description": "Error 400; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "401": {
            "description": "Error 401; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Error 403; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Error 404; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Error 409; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Error 422; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Error 429; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Error 500; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Error 503; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Error 504; sin SQL, tokens ni datos privados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "MyGroups": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "name": {
                  "type": "string"
                },
                "sport": {
                  "type": "string",
                  "nullable": true
                },
                "logo_url": {
                  "type": "string",
                  "nullable": true
                },
                "roles": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "enum": [
                      "ADMIN",
                      "ATHLETE",
                      "GUARDIAN",
                      "COACH"
                    ]
                  },
                  "minItems": 1
                }
              },
              "required": [
                "id",
                "name",
                "sport",
                "logo_url",
                "roles"
              ],
              "additionalProperties": false
            }
          },
          "pagination": {
            "type": "object",
            "properties": {
              "page": {
                "type": "integer",
                "minimum": 1
              },
              "page_size": {
                "type": "integer",
                "minimum": 1,
                "maximum": 100
              },
              "total": {
                "type": "integer",
                "minimum": 0
              }
            },
            "required": [
              "page",
              "page_size",
              "total"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "data",
          "pagination"
        ],
        "additionalProperties": false
      },
...
      "CreateGroup": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 3,
            "maxLength": 80
          },
          "sport": {
            "type": "string",
            "minLength": 2,
            "maxLength": 50
          },
          "description": {
            "type": "string"
          },
          "logo_url": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "sport"
        ],
        "additionalProperties": false
      },
      "GroupCreated": {
        "type": "object",
        "properties": {
          "group_id": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": [
          "group_id"
        ],
        "additionalProperties": false
      },
...
      "GroupSettingsChange": {
        "type": "object",
        "properties": {
          "athletes_can_view_group_stats": {
            "type": "boolean"
          },
          "guardians_can_view_group_stats": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
      "GroupSettings": {
        "type": "object",
        "properties": {
          "settings": {
            "type": "object",
            "properties": {
              "athletes_can_view_group_stats": {
                "type": "boolean"
              },
              "guardians_can_view_group_stats": {
                "type": "boolean"
              }
            },
            "required": [
              "athletes_can_view_group_stats",
              "guardians_can_view_group_stats"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "settings"
        ],
        "additionalProperties": false
      },
...
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "success"
        ],
        "additionalProperties": false
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "minLength": 1
              },
              "message": {
                "type": "string",
                "minLength": 1
              },
              "details": {
                "type": "object",
                "additionalProperties": {}
              }
            },
            "required": [
              "code",
              "message",
              "details"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
    }
  }
}